+++ ports/shared/vc_mobile.cpp
@@ -94,6 +94,31 @@ VcVolume *vc_open (const VcOpenOptions *options, int *error)
 		if (!EncryptionThreadPool::IsRunning ())
 			EncryptionThreadPool::Start ();
 
+		shared_ptr <VolumePassword> protPassword;
+		int protPim = 0;
+		shared_ptr <KeyfileList> protKeyfiles;
+		VolumeProtection::Enum protection = VolumeProtection::None;
+		if (options->read_only)
+			protection = VolumeProtection::ReadOnly;
+		else if (options->protect_hidden)
+		{
+			const char *hp = options->hidden_password ? options->hidden_password : "";
+			size_t hpLen = options->hidden_password_len;
+			if (!hpLen && options->hidden_password)
+				hpLen = strlen (options->hidden_password);
+			if (hpLen == 0 && options->hidden_keyfile_count == 0)
+			{
+				if (error)
+					*error = VC_ERR_ARGUMENT;
+				return nullptr;
+			}
+			protPassword.reset (new VolumePassword (
+				reinterpret_cast <const uint8 *> (hp), hpLen));
+			protPim = options->hidden_pim;
+			protKeyfiles = MakeKeyfilesFrom (options->hidden_keyfiles, options->hidden_keyfile_count);
+			protection = VolumeProtection::HiddenVolumeReadOnly;
+		}
+
 		volume->Open (
 			VolumePath (wstring (options->path, options->path + strlen (options->path))),
 			true,
@@ -102,11 +127,11 @@ VcVolume *vc_open (const VcOpenOptions *options, int *error)
 			shared_ptr <Pkcs5Kdf> (),
 			MakeKeyfiles (options),
 			false,
-			options->read_only ? VolumeProtection::ReadOnly : VolumeProtection::None,
-			shared_ptr <VolumePassword> (),
-			0,
+			protection,
+			protPassword,
+			protPim,
 			shared_ptr <Pkcs5Kdf> (),
-			shared_ptr <KeyfileList> (),
+			protKeyfiles,
 			false,
 			VolumeType::Unknown,
 			options->use_backup_header != 0,
@@ -1035,6 +1060,8 @@ static int fat_writable (VcVolume *volume)
 		return VC_ERR_ARGUMENT;
 	if (volume->read_only)
 		return VC_ERR_UNSUPPORTED;
+	if (volume->volume->IsHiddenVolumeProtectionTriggered ())
+		return VC_ERR_IO;
 	return VC_OK;
 }
 
@@ -2262,13 +2289,24 @@ int vc_volume_info (VcVolume *volume, char *out, size_t out_size)
 	{
 		string cipher = StringConverter::ToSingle (volume->volume->GetEncryptionAlgorithm ()->GetName (true));
 		string kdf = StringConverter::ToSingle (volume->volume->GetPkcs5Kdf ()->GetName ());
-		const char *type = volume->volume->GetType () == VolumeType::Hidden ? "Hidden" : "Normal";
+		const char *type = "Normal";
+		if (volume->volume->GetType () == VolumeType::Hidden)
+			type = "Hidden";
+		else if (volume->volume->GetProtectionType () == VolumeProtection::HiddenVolumeReadOnly)
+			type = "Outer";
+		const char *hiddenProt = "No";
+		if (volume->volume->IsHiddenVolumeProtectionTriggered ())
+			hiddenProt = "Yes (damage prevented!)";
+		else if (volume->volume->GetProtectionType () == VolumeProtection::HiddenVolumeReadOnly)
+			hiddenProt = "Yes";
 		char line[512];
 		snprintf (line, sizeof (line),
-			"Encryption algorithm: %s\nKDF: %s\nVolume type: %s\nSize: %llu bytes\nPIM: %d\nXTS",
+			"Encryption algorithm: %s\nKDF: %s\nVolume type: %s\nSize: %llu bytes\nPIM: %d\nXTS\nHidden Volume Protected: %s\nSector size: %u",
 			cipher.c_str (), kdf.c_str (), type,
 			(unsigned long long) volume->volume->GetSize (),
-			volume->volume->GetPim ());
+			volume->volume->GetPim (),
+			hiddenProt,
+			(unsigned) volume->volume->GetSectorSize ());
 		snprintf (out, out_size, "%s", line);
 		return VC_OK;
 	}
@@ -2276,6 +2314,13 @@ int vc_volume_info (VcVolume *volume, char *out, size_t out_size)
 	{
 		return VC_ERR_FORMAT;
 	}
+}
+
+int vc_protection_triggered (VcVolume *volume)
+{
+	if (!volume || !volume->volume)
+		return 0;
+	return volume->volume->IsHiddenVolumeProtectionTriggered () ? 1 : 0;
 }
 
 int vc_benchmark (char *out, size_t out_size)